+++ client/layouts/default.vue
@@ -10,7 +10,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import NavBar from '@/components/NavBar.vue'
+import NavBar from '@/components/presentational/NavBar.vue'
 
 export default Vue.extend({
   components: {
